Move decorators to @babel/plugin-class-features
