+++ packages/babel-plugin-proposal-decorators/package.json
@@ -16,8 +16,7 @@
   ],
   "dependencies": {
     "@babel/helper-plugin-utils": "^7.0.0",
-    "@babel/helper-replace-supers": "^7.1.0",
-    "@babel/helper-split-export-declaration": "^7.0.0",
+    "@babel/helper-create-class-features-plugin": "^7.2.1",
     "@babel/plugin-syntax-decorators": "^7.2.0"
   },
   "peerDependencies": {
