+++ packages/babel-plugin-proposal-decorators/src/index.js
@@ -1,6 +1,9 @@
 import { declare } from "@babel/helper-plugin-utils";
 import syntaxDecorators from "@babel/plugin-syntax-decorators";
-import visitor from "./transformer";
+import pluginClassFeatures, {
+  enableFeature,
+  FEATURES,
+} from "@babel/plugin-class-features";
 import legacyVisitor from "./transformer-legacy";
 
 export default declare((api, options) => {
@@ -31,14 +34,28 @@ export default declare((api, options) => {
     }
   }
 
+  if (legacy) {
+    return {
+      name: "proposal-decorators",
+      inherits: syntaxDecorators,
+      manipulateOptions({ generatorOpts }) {
+        generatorOpts.decoratorsBeforeExport = decoratorsBeforeExport;
+      },
+      visitor: legacyVisitor,
+    };
+  }
+
   return {
     name: "proposal-decorators",
-    inherits: syntaxDecorators,
+    inherits: pluginClassFeatures,
 
-    manipulateOptions({ generatorOpts }) {
+    manipulateOptions({ generatorOpts, parserOpts }) {
+      parserOpts.plugins.push(["decorators", { decoratorsBeforeExport }]);
       generatorOpts.decoratorsBeforeExport = decoratorsBeforeExport;
     },
 
-    visitor: legacy ? legacyVisitor : visitor,
+    pre() {
+      enableFeature(this.file, FEATURES.decorators, false);
+    },
   };
 });
